fix: expose runtime pool console in visible settings
